requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
paho-mqtt==1.6.1
httpx==0.25.2